Version 1.5.0

+++ src/Jobs/GroupSync.php
@@ -12,6 +12,7 @@ use Herpaderpaldent\Seat\SeatGroups\Exceptions\MissingRefreshTokenException;
 use Herpaderpaldent\Seat\SeatGroups\Models\Seatgroup;
 use Herpaderpaldent\Seat\SeatGroups\Models\SeatgroupLog;
 use Illuminate\Support\Facades\Redis;
+use Seat\Web\Models\Acl\Role;
 use Seat\Web\Models\Group;
 
 class GroupSync extends SeatGroupsJobBase
@@ -43,6 +44,7 @@ class GroupSync extends SeatGroupsJobBase
      */
     public function __construct(Group $group)
     {
+
         $this->group = $group;
         $this->main_character = $group->main_character;
         if (is_null($group->main_character)) {
@@ -64,23 +66,27 @@ class GroupSync extends SeatGroupsJobBase
 
     public function handle()
     {
+
         // in case no main character has been set, throw an exception and abort the process
         if (is_null($this->main_character))
             throw new MissingMainCharacterException($this->group);
 
-        Redis::funnel('seat-groups:jobs.group_sync_' . $this->group->id)->limit(1)->then(function ()
-        {
+        Redis::funnel('seat-groups:jobs.group_sync_' . $this->group->id)->limit(1)->then(function () {
+
             $this->beforeStart();
 
             try {
                 $roles = collect();
                 $group = $this->group;
 
-                //Catch Superuser
+                //Catch superuser permissions
                 foreach ($group->roles as $role) {
-                    if ($role->title === 'Superuser') {
-                        $roles->push($role->id);
+                    foreach ($role->permissions as $permission) {
+                        if ($permission->title === 'superuser') {
+                            $roles->push($role->id);
+                        }
                     }
+
                 }
 
                 Seatgroup::all()->each(function ($seat_group) use ($roles, $group) {
@@ -91,7 +97,7 @@ class GroupSync extends SeatGroupsJobBase
                                 foreach ($seat_group->role as $role) {
                                     $roles->push($role->id);
                                 }
-                                if(! in_array($group->id, $seat_group->group->pluck('id')->toArray())){
+                                if (! in_array($group->id, $seat_group->group->pluck('id')->toArray())) {
                                     // add user_group to seat_group as member if no member yet.
                                     $seat_group->member()->attach($group->id);
                                 }
@@ -107,14 +113,14 @@ class GroupSync extends SeatGroupsJobBase
                                 }
                                 break;
                         }
-                    } elseif(in_array($group->id, $seat_group->group->pluck('id')->toArray())) {
+                    } elseif (in_array($group->id, $seat_group->group->pluck('id')->toArray())) {
                         $seat_group->member()->detach($group->id);
                     }
                 });
 
-                $group->roles()->sync($roles->unique());
+                $sync = $group->roles()->sync($roles->unique());
 
-                $this->onFinish();
+                $this->onFinish($sync);
 
                 logger()->debug('Group has beend synced for ' . $this->main_character->name);
 
@@ -124,8 +130,8 @@ class GroupSync extends SeatGroupsJobBase
 
             }
 
-        }, function ()
-        {
+        }, function () {
+
             logger()->warning('A GroupSync job is already running for ' . $this->main_character->name . ' Removing the job from the queue.');
 
             $this->delete();
@@ -138,25 +144,34 @@ class GroupSync extends SeatGroupsJobBase
 
         foreach ($this->group->users as $user) {
 
+            //If user is deactivated skip the refresh_token check
+            if (! $user->active)
+                continue;
+
             // If a RefreshToken is missing
             if (is_null($user->refresh_token)) {
                 // take away all roles
                 $this->group->roles()->sync([]);
                 Seatgroup::all()->each(function ($seatgroup) {
+
                     $seatgroup->member()->detach($this->group->id);
                 });
 
                 SeatgroupLog::create([
-                    'event' => 'warning',
-                    'message' => sprintf('The RefreshToken of %s is missing, therefore user group of %s (%s) loses all permissions.',
-                        $user->name, $this->main_character->name, $this->group->users->map(function ($user) { return $user->name; })->implode(', ')),
-
+                    'event'   => 'error',
+                    'message' => sprintf('The RefreshToken of %s is missing, therefore user group of %s (%s) loses all permissions.' .
+                        'Ask the owner of this user group to login again with this user, in order to provide a new RefreshToken. ',
+                        $user->name, $this->main_character->name, $this->group->users->map(function ($user) {return $user->name; })->implode(', ')),
                 ]);
 
                 // throw exception
-                throw new MissingRefreshTokenException();
+                $this->fail(new MissingRefreshTokenException($user));
             }
         }
+
+        // If deactivated user is alone in a group delete this job
+        if (! $this->group->users->first()->active && $this->group->users->count() === 1)
+            $this->delete();
     }
 
     public function onFail($exception)
@@ -165,21 +180,46 @@ class GroupSync extends SeatGroupsJobBase
         report($exception);
 
         SeatgroupLog::create([
-            'event' => 'error',
+            'event'   => 'error',
             'message' => sprintf('An error occurred while syncing user group of %s (%s). Please check the logs.',
-                $this->main_character->name, $this->group->users->map(function ($user) { return $user->name; })->implode(', ')),
-
+                $this->main_character->name, $this->group->users->map(function ($user) {return $user->name; })->implode(', ')),
         ]);
+
+        $this->fail($exception);
 
     }
 
-    public function onFinish()
+    public function onFinish($sync)
     {
-        SeatgroupLog::create([
-            'event' => 'success',
-            'message' => sprintf('The user group of %s (%s) has successfully been synced.',
-                $this->main_character->name, $this->group->users->map(function ($user) { return $user->name; })->implode(', ')),
 
-        ]);
+        if (! empty($sync['attached'])) {
+
+            SeatgroupLog::create([
+                'event'   => 'attached',
+                'message' => sprintf('The user group of %s (%s) has successfully been attached to the following roles: %s.',
+                    $this->main_character->name,
+                    $this->group->users->map(function ($user) {
+
+                        return $user->name;
+                    })->implode(', '),
+                    Role::whereIn('id', $sync['attached'])->pluck('title')->implode(', ')
+                ),
+            ]);
+        }
+
+        if (! empty($sync['detached'])) {
+
+            SeatgroupLog::create([
+                'event'   => 'detached',
+                'message' => sprintf('The user group of %s (%s) has been detached from the following roles: %s.',
+                    $this->main_character->name,
+                    $this->group->users->map(function ($user) {
+
+                        return $user->name;
+                    })->implode(', '),
+                    Role::whereIn('id', $sync['detached'])->pluck('title')->implode(', ')
+                ),
+            ]);
+        }
     }
 }
